+++ crates/client/rpc/src/methods/read/get_class_at.rs
@@ -28,7 +28,7 @@ use crate::methods::trace::utils::block_number_by_id;
 /// * `BLOCK_NOT_FOUND` - If the specified block does not exist in the blockchain.
 /// * `CONTRACT_NOT_FOUND` - If the specified contract address does not exist.
 pub fn get_class_at(block_id: BlockId, contract_address: FieldElement) -> RpcResult<ContractClass> {
-    let block_number = block_number_by_id(block_id);
+    let block_number = block_number_by_id(block_id)?;
     let key = ContractAddress(PatriciaKey(StarkFelt(contract_address.to_bytes_be())));
 
     let class_hash = match storage_handler::contract_data().get_class_hash_at(&key, block_number) {