+++ crates/client/rpc/src/methods/read/get_nonce.rs
@@ -28,7 +28,7 @@ use crate::Felt;
 pub fn get_nonce(block_id: BlockId, contract_address: FieldElement) -> RpcResult<Felt> {
     let key = ContractAddress(PatriciaKey(StarkFelt(contract_address.to_bytes_be())));
 
-    let block_number = block_number_by_id(block_id);
+    let block_number = block_number_by_id(block_id)?;
     match storage_handler::contract_data().get_nonce_at(&key, block_number) {
         Err(e) => {
             log::error!("Failed to get nonce: {e}");