+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next release
 
+- refactor: using const and OnceCell instead of lazy_static
+- refactor: remove crate mp-storage
 - feat(infra): corrected dockerfile + docker-compose
 - fix(rpc): error handling
 - fix(lib): updated core libs to match oss